refactor: :recycle: generalise InvalidResourcePropertiesError to InvalidPropertiesError
